Organize Components and create css files
