+++ client/src/components/feed/feed.css
@@ -1,0 +1,14 @@
+.feed-item-img {
+  width: "100%";
+  height: "12vw";
+  object-fit: "cover";
+}
+
+.reading-font {
+  font-family: "Lato sans-serif";
+}
+
+a {
+  text-decoration: none;
+  color: black;
+}
