+++ client/src/components/note/note.css
@@ -1,0 +1,8 @@
+.note-feature-img {
+  margin: 20px 0;
+  width: 100%;
+}
+
+.note {
+  font-family: "Lato sans-serif";
+}
